refactor(data table manager): refactor manager provider
