+++ packages/components/data-table-manager/README.md
@@ -72,6 +72,7 @@ import DataTableManager, {
   DataTableManagerProvider,
 } from '@commercetools-uikit/data-table-manager';
 import DataTable from '@commercetools-uikit/data-table';
+import SearchTextInput from '@commercetools-uikit/search-text-input';
 
 const rows = [
   { id: 'parasite', title: 'Parasite', country: 'South Korea' },
@@ -90,9 +91,14 @@ const SomeOtherComponent = () => {
 
 const Example = () => (
   <DataTableManagerProvider columns={columns}>
-    <DataTableManager />
-    <SomeOtherComponent />
-    <DataTable rows={rows} />
+    <header>
+      <DataTableManager />
+      <SearchTextInput />
+    </header>
+    <main>
+      <SomeOtherComponent />
+      <DataTable rows={rows} />
+    </main>
   </DataTableManagerProvider>
 );
