+++ packages/components/data-table-manager/src/export-types.ts
@@ -1,2 +1,8 @@
-export type { TRow, TColumnProps, TDataTableManagerProps } from './types';
+export type {
+  TRow,
+  TColumnProps,
+  TDataTableManagerProps,
+  TColumnData,
+  TDataTableSettingsProps,
+} from './types';
 export type { TDataTableManagerContext } from './data-table-manager-provider';
